requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
orjson==3.9.10
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.8.0
ulid-py==1.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
